requirements.txt
selenium == 4.39.0
requests == 2.32.5
pytest == 9.0.2
allure-pytest == 2.15.2